Refactor JsonlDecoder and create new LogbackFormatter class.
